+++ resources/category.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const Ciceksepeti = require('..');
 const axios = require('axios');
 
 /**
@@ -16,7 +17,6 @@ function Category(ciceksepeti) {
 /**
  * Returns a list of Categories.
  *
- * @param {Object} params Query parameters
  * @return {Promise} Promise that resolves with the result
  * @public
  */
@@ -25,6 +25,43 @@ Category.prototype.list = async function list() {
         + '/api'
         + '/' + this.ciceksepeti.options.apiVersion
         + '/Categories';
+
+    let config = {
+        method: 'get',
+        maxBodyLength: Infinity,
+        url: url,
+        headers: this.ciceksepeti.baseHeaders,
+        maxRedirects: 0,
+    };
+
+    return axios(config)
+        .then(function (response) {
+            return response.data;
+        })
+        .catch(function (error) {
+            throw new Error(error.response.data['Message']);
+        });
+};
+
+/**
+ * Returns attributes of a Category.
+ *
+ * @param {Object} params Query parameters
+ * @return {Promise} Promise that resolves with the result
+ * @public
+ */
+Category.prototype.attributes = async function attributes(params) {
+    params = params || {};
+    if (!params.id) {
+        throw new Error('Category ID (id) is required.');
+    }
+
+    let url = this.ciceksepeti.baseUrl.protocol + '//' + this.ciceksepeti.baseUrl.hostname
+        + '/api'
+        + '/' + this.ciceksepeti.options.apiVersion
+        + '/Categories'
+        + '/' + params.id
+        + '/attributes';
 
     let config = {
         method: 'get',