+++ resources/index.js
@@ -4,6 +4,7 @@ const map = {
     product: 'product',
     order: 'order',
     category: 'category',
+    sellerquestion: 'sellerquestion',
 };
 
 /**